test: SlackInteractionInbox JDBC Fixture 추출
