+++ src/test/java/com/slack/bot/application/interaction/box/in/SlackInteractionInboxProcessorTest.java
@@ -20,17 +20,12 @@ import com.slack.bot.application.worker.PollingHintTarget;
 import com.slack.bot.domain.reservation.ReservationStatus;
 import com.slack.bot.domain.reservation.ReviewReservation;
 import com.slack.bot.domain.reservation.repository.ReviewReservationRepository;
-import com.slack.bot.infrastructure.common.BoxEventTime;
-import com.slack.bot.infrastructure.common.BoxFailureSnapshot;
-import com.slack.bot.infrastructure.common.BoxProcessingLease;
 import com.slack.bot.infrastructure.interaction.box.SlackInteractionFailureType;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInbox;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxHistory;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxStatus;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxType;
 import com.slack.bot.infrastructure.interaction.box.in.repository.SlackInteractionInboxRepository;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
@@ -41,6 +36,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -52,6 +48,7 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.support.TransactionTemplate;
+import com.slack.bot.support.SlackInteractionInboxJdbcFixture;
 
 @IntegrationTest
 @MockitoSpyBean(types = PollingHintPublisher.class)
@@ -85,6 +82,13 @@ class SlackInteractionInboxProcessorTest {
 
     @Autowired
     PlatformTransactionManager transactionManager;
+
+    SlackInteractionInboxJdbcFixture slackInteractionInboxJdbcFixture;
+
+    @BeforeEach
+    void setUp() {
+        slackInteractionInboxJdbcFixture = new SlackInteractionInboxJdbcFixture(namedParameterJdbcTemplate);
+    }
 
     @Test
     @Sql(scripts = {
@@ -412,152 +416,19 @@ class SlackInteractionInboxProcessorTest {
     }
 
     private List<SlackInteractionInboxHistory> historiesOf(Long inboxId) {
-        return namedParameterJdbcTemplate.query(
-                """
-                SELECT id,
-                       inbox_id,
-                       processing_attempt,
-                       status,
-                       completed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox_history
-                WHERE inbox_id = :inboxId
-                ORDER BY processing_attempt ASC
-                """,
-                new MapSqlParameterSource().addValue("inboxId", inboxId),
-                (resultSet, rowNum) -> mapHistory(resultSet)
-        );
+        return slackInteractionInboxJdbcFixture.historiesOf(inboxId);
     }
 
     private List<SlackInteractionInbox> findAllInboxes() {
-        return namedParameterJdbcTemplate.query(
-                """
-                SELECT id,
-                       interaction_type,
-                       idempotency_key,
-                       payload_json,
-                       status,
-                       processing_attempt,
-                       processing_started_at,
-                       processed_at,
-                       failed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox
-                ORDER BY id ASC
-                """,
-                (resultSet, rowNum) -> mapInbox(resultSet)
-        );
+        return slackInteractionInboxJdbcFixture.findAllInboxes();
     }
 
     private List<SlackInteractionInbox> findInboxesByIds(List<Long> inboxIds) {
-        return namedParameterJdbcTemplate.query(
-                """
-                SELECT id,
-                       interaction_type,
-                       idempotency_key,
-                       payload_json,
-                       status,
-                       processing_attempt,
-                       processing_started_at,
-                       processed_at,
-                       failed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox
-                WHERE id IN (:inboxIds)
-                ORDER BY id ASC
-                """,
-                new MapSqlParameterSource().addValue("inboxIds", inboxIds),
-                (resultSet, rowNum) -> mapInbox(resultSet)
-        );
+        return slackInteractionInboxJdbcFixture.findInboxesByIds(inboxIds);
     }
 
     private List<SlackInteractionInboxHistory> findHistoriesByInboxIds(List<Long> inboxIds) {
-        return namedParameterJdbcTemplate.query(
-                """
-                SELECT id,
-                       inbox_id,
-                       processing_attempt,
-                       status,
-                       completed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox_history
-                WHERE inbox_id IN (:inboxIds)
-                ORDER BY inbox_id ASC, processing_attempt ASC
-                """,
-                new MapSqlParameterSource().addValue("inboxIds", inboxIds),
-                (resultSet, rowNum) -> mapHistory(resultSet)
-        );
-    }
-
-    private SlackInteractionInbox mapInbox(ResultSet resultSet) throws SQLException {
-        return SlackInteractionInbox.rehydrate(
-                resultSet.getLong("id"),
-                SlackInteractionInboxType.valueOf(resultSet.getString("interaction_type")),
-                resultSet.getString("idempotency_key"),
-                resultSet.getString("payload_json"),
-                SlackInteractionInboxStatus.valueOf(resultSet.getString("status")),
-                resultSet.getInt("processing_attempt"),
-                toProcessingLease(resultSet),
-                toProcessedTime(resultSet),
-                toFailedTime(resultSet),
-                toFailure(resultSet)
-        );
-    }
-
-    private SlackInteractionInboxHistory mapHistory(ResultSet resultSet) throws SQLException {
-        return SlackInteractionInboxHistory.rehydrate(
-                resultSet.getLong("id"),
-                resultSet.getLong("inbox_id"),
-                resultSet.getInt("processing_attempt"),
-                SlackInteractionInboxStatus.valueOf(resultSet.getString("status")),
-                resultSet.getTimestamp("completed_at").toInstant(),
-                toFailure(resultSet)
-        );
-    }
-
-    private BoxProcessingLease toProcessingLease(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("processing_started_at") == null) {
-            return BoxProcessingLease.idle();
-        }
-
-        return BoxProcessingLease.claimed(resultSet.getTimestamp("processing_started_at").toInstant());
-    }
-
-    private BoxEventTime toProcessedTime(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("processed_at") == null) {
-            return BoxEventTime.absent();
-        }
-
-        return BoxEventTime.present(resultSet.getTimestamp("processed_at").toInstant());
-    }
-
-    private BoxEventTime toFailedTime(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("failed_at") == null) {
-            return BoxEventTime.absent();
-        }
-
-        return BoxEventTime.present(resultSet.getTimestamp("failed_at").toInstant());
-    }
-
-    private BoxFailureSnapshot<SlackInteractionFailureType> toFailure(ResultSet resultSet) throws SQLException {
-        String failureReason = resultSet.getString("failure_reason");
-        String failureType = resultSet.getString("failure_type");
-
-        if (failureReason == null && failureType == null) {
-            return BoxFailureSnapshot.absent();
-        }
-        if (failureReason == null || failureType == null) {
-            throw new IllegalStateException("failure_reason과 failure_type은 항상 함께 세팅되어야 합니다.");
-        }
-
-        return BoxFailureSnapshot.present(
-                failureReason,
-                SlackInteractionFailureType.valueOf(failureType)
-        );
+        return slackInteractionInboxJdbcFixture.findHistoriesByInboxIds(inboxIds);
     }
 
     private void awaitLatch(CountDownLatch latch) {
