+++ src/test/java/com/slack/bot/infrastructure/interaction/box/persistence/in/SlackInteractionInboxRepositoryAdapterTest.java
@@ -5,19 +5,15 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.slack.bot.application.IntegrationTest;
-import com.slack.bot.infrastructure.common.BoxEventTime;
-import com.slack.bot.infrastructure.common.BoxFailureSnapshot;
-import com.slack.bot.infrastructure.common.BoxProcessingLease;
 import com.slack.bot.infrastructure.interaction.box.SlackInteractionFailureType;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInbox;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxHistory;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxStatus;
 import com.slack.bot.infrastructure.interaction.box.in.SlackInteractionInboxType;
 import com.slack.bot.infrastructure.interaction.box.in.repository.SlackInteractionInboxRepository;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.time.Instant;
 import java.util.List;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -25,6 +21,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import com.slack.bot.support.SlackInteractionInboxJdbcFixture;
 
 @IntegrationTest
 @SuppressWarnings("NonAsciiCharacters")
@@ -36,6 +33,13 @@ class SlackInteractionInboxRepositoryAdapterTest {
 
     @Autowired
     NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
+    SlackInteractionInboxJdbcFixture slackInteractionInboxJdbcFixture;
+
+    @BeforeEach
+    void setUp() {
+        slackInteractionInboxJdbcFixture = new SlackInteractionInboxJdbcFixture(namedParameterJdbcTemplate);
+    }
 
     @Test
     void recoverTimeoutProcessing은_interactionType이_null이면_예외를_던진다() {
@@ -334,44 +338,11 @@ class SlackInteractionInboxRepositoryAdapterTest {
     }
 
     private List<SlackInteractionInboxHistory> historiesOf(Long inboxId) {
-        return namedParameterJdbcTemplate.query(
-                """
-                SELECT id,
-                       inbox_id,
-                       processing_attempt,
-                       status,
-                       completed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox_history
-                WHERE inbox_id = :inboxId
-                ORDER BY processing_attempt ASC
-                """,
-                new MapSqlParameterSource().addValue("inboxId", inboxId),
-                (resultSet, rowNum) -> mapHistory(resultSet)
-        );
+        return slackInteractionInboxJdbcFixture.historiesOf(inboxId);
     }
 
     private SlackInteractionInbox findInboxByIdempotencyKey(String idempotencyKey) {
-        return namedParameterJdbcTemplate.queryForObject(
-                """
-                SELECT id,
-                       interaction_type,
-                       idempotency_key,
-                       payload_json,
-                       status,
-                       processing_attempt,
-                       processing_started_at,
-                       processed_at,
-                       failed_at,
-                       failure_reason,
-                       failure_type
-                FROM slack_interaction_inbox
-                WHERE idempotency_key = :idempotencyKey
-                """,
-                new MapSqlParameterSource().addValue("idempotencyKey", idempotencyKey),
-                (resultSet, rowNum) -> mapInbox(resultSet)
-        );
+        return slackInteractionInboxJdbcFixture.findInboxByIdempotencyKey(idempotencyKey);
     }
 
     private int countInboxByIdempotencyKey(String idempotencyKey) {
@@ -408,81 +379,4 @@ class SlackInteractionInboxRepositoryAdapterTest {
         );
     }
 
-    private SlackInteractionInbox mapInbox(ResultSet resultSet) throws SQLException {
-        BoxProcessingLease processingLease = toProcessingLease(resultSet);
-        BoxEventTime processedTime = toProcessedTime(resultSet);
-        BoxEventTime failedTime = toFailedTime(resultSet);
-        BoxFailureSnapshot<SlackInteractionFailureType> failure = toFailure(resultSet);
-
-        return SlackInteractionInbox.rehydrate(
-                resultSet.getLong("id"),
-                SlackInteractionInboxType.valueOf(resultSet.getString("interaction_type")),
-                resultSet.getString("idempotency_key"),
-                resultSet.getString("payload_json"),
-                SlackInteractionInboxStatus.valueOf(resultSet.getString("status")),
-                resultSet.getInt("processing_attempt"),
-                processingLease,
-                processedTime,
-                failedTime,
-                failure
-        );
-    }
-
-    private SlackInteractionInboxHistory mapHistory(ResultSet resultSet) throws SQLException {
-        return SlackInteractionInboxHistory.rehydrate(
-                resultSet.getLong("id"),
-                resultSet.getLong("inbox_id"),
-                resultSet.getInt("processing_attempt"),
-                SlackInteractionInboxStatus.valueOf(resultSet.getString("status")),
-                resultSet.getTimestamp("completed_at").toInstant(),
-                toFailure(resultSet)
-        );
-    }
-
-    private BoxProcessingLease toProcessingLease(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("processing_started_at") == null) {
-            return BoxProcessingLease.idle();
-        }
-
-        return BoxProcessingLease.claimed(
-                resultSet.getTimestamp("processing_started_at").toInstant()
-        );
-    }
-
-    private BoxEventTime toProcessedTime(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("processed_at") == null) {
-            return BoxEventTime.absent();
-        }
-
-        return BoxEventTime.present(
-                resultSet.getTimestamp("processed_at").toInstant()
-        );
-    }
-
-    private BoxEventTime toFailedTime(ResultSet resultSet) throws SQLException {
-        if (resultSet.getTimestamp("failed_at") == null) {
-            return BoxEventTime.absent();
-        }
-
-        return BoxEventTime.present(
-                resultSet.getTimestamp("failed_at").toInstant()
-        );
-    }
-
-    private BoxFailureSnapshot<SlackInteractionFailureType> toFailure(ResultSet resultSet) throws SQLException {
-        String failureReason = resultSet.getString("failure_reason");
-        String failureType = resultSet.getString("failure_type");
-
-        if (failureReason == null && failureType == null) {
-            return BoxFailureSnapshot.absent();
-        }
-        if (failureReason == null || failureType == null) {
-            throw new IllegalStateException("failure_reason과 failure_type은 항상 함께 세팅되어야 합니다.");
-        }
-
-        return BoxFailureSnapshot.present(
-                failureReason,
-                SlackInteractionFailureType.valueOf(failureType)
-        );
-    }
 }
